+++ Case.py
@@ -20,7 +20,7 @@ from z3 import And, Not, sat, unsat, unknown, is_true
 
 from Idp.Expression import Brackets, AUnary, TRUE, FALSE, AppliedSymbol, Variable, AConjunction, ADisjunction, AComparison
 from Solver import mk_solver
-from Structure_ import json_to_literals, Equality, Assignment, Term, Status
+from Structure_ import json_to_literals, Assignment, Status, str_to_IDP
 from utils import *
 
 # Types
@@ -59,20 +59,21 @@ class Case:
                 or any(s in self.expanded_symbols for s in GuiLine.unknown_symbols().keys())
 
         # initialize .assignments
-        self.assignments = {s.code : Assignment(s, None, Status.UNKNOWN) for s in self.idp.subtences.values()}
+        self.assignments = {s.code : Assignment(s.copy(), None, Status.UNKNOWN) for s in self.idp.subtences.values()}
         self.assignments.update({ atom.code : ass for atom, ass in self.given.items() })
 
         # find immediate universals
         for i, c in enumerate(self.idp.theory.constraints):
-            u = c.as_substitutions(self)
+            c = c.copy()
+            u = c.implicants()
             if u:
                 for sentence, truth in u:
-                    ass = Assignment(sentence, truth==TRUE, Status.UNKNOWN)
-                    ass.update(None, None, Status.UNIVERSAL, self)
-            else:
-                self.simplified.append(c)
+                    ass = Assignment(sentence, truth, Status.UNKNOWN)
+                    status = Status.ENV_UNIV if not ass.sentence.has_environmental(False) else Status.UNIVERSAL
+                    ass.update(None, None, status, self)
+            self.simplified.append(c)
 
-        self.assignments.update({ k : Term(Equality(t, None), None, Status.UNKNOWN) 
+        self.assignments.update({ k : Assignment(t.copy(), None, Status.UNKNOWN) 
             for k, t in idp.vocabulary.terms.items()
             if k not in self.assignments })
 
@@ -81,24 +82,29 @@ class Case:
             self.full_propagate(all_=False)
         self.full_propagate(all_=True) # now consider all facts and theories
 
+        # determine relevant assignemnts
+        relevant_symbols, relevant_subtences = self.get_relevant_subtences(all_=True)
+
+        for k, l in self.assignments.items():
+            symbols = list(l.sentence.unknown_symbols().keys())
+            has_relevant_symbol = any(s in relevant_symbols for s in symbols)
+            if k in relevant_subtences and symbols and has_relevant_symbol:
+                l.relevant = True
         if DEBUG: assert invariant == ".".join(str(e) for e in self.idp.theory.constraints)
 
     def __str__(self) -> str:
-        return (f"Type:        {indented}{indented.join(repr(d) for d in self.typeConstraints.translated)}{nl}"
+        return (f"Type:        {indented}{indented.join(repr(d) for d in self.typeConstraints.translate(self.idp))}{nl}"
                 f"Definitions: {indented}{indented.join(repr(d) for d in self.definitions)}{nl}"
-                f"Universals:  {indented}{indented.join(repr(c) for c in self.assignments.values() if c.status == Status.UNIVERSAL)}{nl}"
+                f"Universals:  {indented}{indented.join(repr(c) for c in self.assignments.values() if c.status in [Status.UNIVERSAL, Status.ENV_UNIV])}{nl}"
                 f"Consequences:{indented}{indented.join(repr(c) for c in self.assignments.values() if c.status in [Status.CONSEQUENCE, Status.ENV_CONSQ])}{nl}"
                 f"Simplified:  {indented}{indented.join(str(c)  for c in self.simplified)}{nl}"
-                f"Irrelevant:  {indented}{indented.join(repr(c) for c in self.assignments.values() if not c.relevant and type(c) != Term)}{nl}"
+                f"Irrelevant:  {indented}{indented.join(repr(c) for c in self.assignments.values() if not c.relevant)}{nl}"
         )
         
     def get_relevant_subtences(self, all_: bool) -> Tuple[Dict[str, SymbolDeclaration], Dict[str, Expression]]:
         #TODO performance.  This method is called many times !  use expr.contains(expr, symbols)
         constraints = ( self.simplified
-            + list(self.idp.goal.subtences().values())
-            + [l.sentence for k, l in self.assignments.items() 
-                    if l.truth is not None and (all_ or l.is_environmental) 
-                    and not type(l) == Term])
+            + list(self.idp.goal.subtences().values()))
 
         # determine relevant symbols (including defined ones)
         symbols = mergeDicts( e.unknown_symbols() for e in constraints )
@@ -119,21 +125,13 @@ class Case:
 
         # simplify all using given and universals
         to_propagate = list(l for l in self.assignments.values() 
-            if l.truth is not None and (all_ or l.is_environmental))
+            if l.truth is not None)
         self.propagate(to_propagate, all_)
 
         Log(f"{nl}Z3 propagation ********************************")
-        # determine relevant assignemnts
-        if all_:
-            relevant_symbols, relevant_subtences = self.get_relevant_subtences(all_=True)
 
-            for k, l in self.assignments.items():
-                symbols = list(l.sentence.unknown_symbols().keys())
-                has_relevant_symbol = any(s in relevant_symbols for s in symbols)
-                if k in relevant_subtences and symbols and has_relevant_symbol:
-                    l.relevant = True
-
-        solver, _, _ = mk_solver(self.translate(all_), {})
+        theory = self.translate(all_)
+        solver, _, _ = mk_solver(theory, {})
         result = solver.check()
         if result == sat:
             todo = self.assignments.keys()
@@ -142,7 +140,7 @@ class Case:
             for key in todo:
                 l = self.assignments[key]
                 if ( l.truth is None
-                and (all_ or l.is_environmental)
+                and (all_ or not l.sentence.has_environmental(False))
                 # and key in self.get_relevant_subtences(all_) 
                 and self.GUILines[key].is_visible):
                     atom = l.sentence
@@ -152,23 +150,18 @@ class Case:
                     if res1 == sat:
                         val1 = solver.model().eval(atom.reified())
                         if str(val1) != str(atom.reified()): # if not irrelevant
+                        
+                            val = str_to_IDP(self.idp, str(val1))
+                            assert val.translate() == val1, str(val) + " is not the same as " + str(val1)
+
                             solver.push()
                             solver.add(Not(atom.reified()==val1))
                             res2 = solver.check()
                             solver.pop()
 
                             if res2 == unsat:
-                                if type(l) == Term:
-                                    atom = cast(Equality, atom)
-                                    # need to convert Term into Assignment
-                                    if atom.variable.decl.out.code == 'bool':
-                                        ass = Assignment(atom.variable, is_true(val1), CONSQ)
-                                    else:
-                                        ass = Assignment(Equality(atom.variable, val1), True, CONSQ)
-                                    ass.relevant = True
-                                    self.assignments[key] = ass
-                                else:
-                                    ass = l.update(None, is_true(val1), CONSQ, self)
+                                ass = l.update(atom, val, CONSQ, self)
+                                #TODO ass.relevant = True
                                 self.propagate([ass], all_)
                             elif res2 == unknown:
                                 res1 = unknown
@@ -192,57 +185,58 @@ class Case:
                 for constraint in self.simplified:
                     consequences = []
                     new_constraint = constraint.substitute(old, new, consequences)
-                    consequences.extend(new_constraint.as_substitutions(self))
+                    consequences.extend(new_constraint.implicants())
                     if consequences:
                         for sentence, truth in consequences:
-                            old_ass = self.assignments[sentence.code]
-                            if old_ass.truth is None:
-                                if (all_ or old_ass.is_environmental):
-                                    new_ass = old_ass.update(None, truth==TRUE, CONSQ, self)
-                                    to_propagate.append(new_ass)
-                            elif old_ass.truth != (truth==TRUE):
-                                # test: theory{ x=4. x=5. }
-                                self.simplified = cast(List[Expression], [FALSE]) # inconsistent !
-                                return
-                        if not any(new_constraint == e for (e,t) in consequences):
-                            new_simplified.append(new_constraint)
-                    elif not new_constraint == TRUE:
-                        new_simplified.append(new_constraint)
+                            if sentence.code in self.assignments:
+                                old_ass = self.assignments[sentence.code]
+                                if old_ass.truth is None:
+                                    if (all_ or not constraint.has_environmental(False)):
+                                        new_ass = old_ass.update(sentence, truth, CONSQ, self)
+                                        to_propagate.append(new_ass)
+                                        new_constraint = new_constraint.substitute(sentence, truth)
+                                elif old_ass.truth != truth:
+                                    # test: theory{ x=4. x=5. }
+                                    self.simplified = cast(List[Expression], [FALSE]) # inconsistent !
+                                    return
+                            else:
+                                print("not found", str(sentence))
+                    new_simplified.append(new_constraint)
 
                 self.simplified = new_simplified
 
                 # simplify assignments
+                # e.g. 'Sides=4' becomes false when Sides becomes 3
                 for old_ass in self.assignments.values():
-                    if old_ass.sentence != ass.sentence and (all_ or old_ass.is_environmental):
-                        new_constraint = old_ass.sentence.substitute(old, new)
-                        if new_constraint != old_ass.sentence: # changed !
-                            if type(old_ass) == Term: # value of term was not known
-                                old_ass = cast(Term, old_ass)
-                                new_ass = old_ass.assign(new_constraint.value, self, CONSQ)
-                                to_propagate.append(new_ass)
-                            elif new_constraint in [TRUE, FALSE]:
-                                if old_ass.truth is None: # value of proposition was not known
-                                    new_ass = old_ass.update(new_constraint, 
-                                        (new_constraint == TRUE), CONSQ, self)
-                                    to_propagate.append(new_ass)
-                                elif (new_constraint==TRUE  and not old_ass.truth) \
-                                or   (new_constraint==FALSE and     old_ass.truth):
-                                    #TODO test case ?
-                                    self.simplified = cast(List[Expression], [FALSE]) # inconsistent
-                                    return
-                                else:
-                                    pass # no change
-                            else:
-                                old_ass.update(new_constraint, None, None, self)
+                    before = str(old_ass.sentence)
+                    new_constraint = old_ass.sentence.substitute(old, new)
+                    if before != str(new_constraint) \
+                    and (all_ or not new_constraint.has_environmental(False)): # is purely environmental
+                        if new_constraint.code in self.assignments \
+                        and self.assignments[new_constraint.code].truth is not None: # e.g. O(M) becomes O(e2)
+                            new_constraint = self.assignments[new_constraint.code].sentence
+                        if new_constraint.value is None: # not reduced to ground
+                            old_ass.update(new_constraint, None, None, self)
+                        elif old_ass.truth is not None: # has a value already
+                            if not old_ass.truth == new_constraint.value: # different !
+                                self.simplified = cast(List[Expression], [FALSE]) # inconsistent
+                                return
+                            else: # no change
+                                pass
+                        else: # accept new value
+                            new_ass = old_ass.update(new_constraint, 
+                                new_constraint.value, CONSQ, self)
+                            to_propagate.append(new_ass)
+                            
 
     def translate(self, all_: bool = True) -> BoolRef:
         self.translated = And(
-            self.typeConstraints.translated
+            self.typeConstraints.translate(self.idp)
             + sum((d.translate(self.idp) for d in self.definitions), [])
             + [l.translate() for k, l in self.assignments.items() 
-                    if l.truth is not None and (all_ or l.is_environmental) 
-                    and not type(l) == Term]
-            + [c.translate() for c in self.simplified]
+                    if l.truth is not None and (all_ or not l.sentence.has_environmental(False))]
+            + [c.translate() for c in self.simplified
+                    if all_ or not c.has_environmental(False)]
             )
         return self.translated