+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
 ### API
+* add ENV_UNIV for universal environmental atoms
 * add English annotations to symbols in vocabulary
 * fix bug with Fresh_Variable